+++ greskell/ChangeLog.md
@@ -9,6 +9,7 @@
 * Add the following functions.
   * `gFlatMap'`
   * `gCoalesce`
+  * `gIterate`
 
 ### Extra module